Quick (preference) style change; "fix" for iOS Embed Link
Although links now work in the Embed body on iOS, wrapping a link in
ticks (`) is not clickable on Discord iOS (TestFlight). Also added >>>
to the output of Eval’s overflow result string, so it can appear on a
newline (might just be the CSS theme I use, but I like it anyways).